requirements.txt
streamlit==1.32.0
spotipy==2.23.0
pandas==2.2.0
matplotlib==3.8.2
python-dotenv==1.0.1 